Pushing Code

+++ Java_to_do/src/java100days/OmHambireDay10.java
@@ -1,5 +1,5 @@
 package java100days;
-//Get a number from user for which you need to fin the factorial, then calculate the factorial and give it as 
+//Get a number from user for which you need to find the factorial, then calculate the factorial and give it as 
 //the output.
 import java.util.*;
 public class OmHambireDay10 
@@ -7,15 +7,15 @@ public class OmHambireDay10
 	public static void main(String[] args)
 	{
 		Scanner sn=new Scanner(System.in);
-		int num=sn.nextInt();
+		long num=sn.nextInt();
 		if (num<0)
 		{
 			System.out.println("please enter positive number");
 		}
 		else 
 		{
-		int fact=1;
-		for(int i=0;i<num;i++)
+		long fact=1;
+		for(long i=0;i<num;i++)
 		{
 			fact=fact*(i+1);
 		}
